Move validateShippingCountry util out of main file
